+++ src/pinyin.h
@@ -88,9 +88,9 @@ bool pinyin_get_candidates(pinyin_instance_t * instance,
                            size_t offset,
                            TokenVector candidates);
 
-guint8 pinyin_choose_candidate(pinyin_instance_t * instance,
-                               size_t offset,
-                               phrase_token_t token);
+int pinyin_choose_candidate(pinyin_instance_t * instance,
+                            size_t offset,
+                            phrase_token_t token);
 
 bool pinyin_clear_constraint(pinyin_instance_t * instance,
                              size_t offset);
